feat: add animations for moving players and enemies
Added fog-of-war style visibility for each character to have some memory of what they've seen.
Removed the action timer bar since it's kinda ugly and I didn't want to figure out how to make it work with the animations.
Increased the color of the background so edges of the world are more visible.
Updated version of android libraries, maybe?