fix: player doesn't stop when interacting with an object at its target location
This is are reversion. It was intended to stop the player from swapping items back and forth. However, that was actually a feature. The logic broke trying to dig an adjacent rock.
Renamed "player" to "survivor" for no good reason.
Changed game creation options to buttons for enemy and item count instead of sliders. Those sliders were annoying to use and options are more descriptive.
Also updated the world size options to be more creative.