feat: make flares better
Flares now light up an approximately circular area for all survivors instead of just the one who used it having unlimited vision.
Blocks that the survivors have seen are updated for all survivors instead of just the currently active survivor.
Increased iteration limit for pathing to improve it a bit.
Increased the default path cost of unseen blocks to 5 so pathing doesn't prefer unseen blocks.
The edges of the world are now black, non-visible blocks are dark gray, and the ground is a dark green.
Visibility code has been updated so there's less double negative silliness.